// web/src/api/iot.ts
export interface IotRealtimeValue {
  tagCode: string
  value: number
  valueText: string
  dataType: string
  unit: string
  quality: string
  timestamp: string
  sourceTime: string
  deviceCode: string
  areaCode: string
  sourceType: string
}